fixed issues incorrect access request params

+++ bottle/code/teachme_api/teachme_api.py
@@ -1,6 +1,6 @@
 from random import randint
 from mongokit import Connection, Document
-from bottle import get, post, put, delete, run
+from bottle import get, post, put, delete, run, request
 
 MONGODB_HOST = 'localhost'
 MONGODB_PORT = 27017
@@ -36,25 +36,26 @@ def show_topic(id):
 def create_topic():
 	topic = connection.Topic()
 	topic._id = randint(1, 1000000)
-	topic.name = request.form['name']
-	topic.description = request.form['description']
-	topic.presenter = request.form['presenter']
+	topic.name = unicode(request.params['name'])
+	topic.description = unicode(request.params['description'])
+	topic.presenter = unicode(request.params['presenter'])
 	topic.save()
 	return topic
 
 @put('/topics/:id')
 def update_topic(id):
 	topic = connection.Topic.find_one({'_id': int(id)})
-	topic.name = request.form['name']
-	topic.description = request.form['description']
-	topic.presenter = request.form['presenter']
+	topic.name = unicode(request.params['name'])
+	topic.description = unicode(request.params['description'])
+	topic.presenter = unicode(request.params['presenter'])
 	topic.save()
 	return topic
 
 @delete('/topics/:id')
 def destroy_topic(id):
 	topic = connection.Topic.find_one({'_id': int(id)})
-	topic.delete()
-	return True
+	if topic is not None:
+		topic.delete()
+	return {'status':'success'}
 
 run(host='localhost', port=8080, reloader=True)
